+++ packages/components/organisms/f-checkout/CHANGELOG.md
@@ -3,6 +3,14 @@
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+v0.52.0
+-------------------------------
+*February 9, 2021*
+
+### Added
+- Utilised `vue-svg-loader` for error page SVG
+- Added `vue-svg-loader` to Webpack config
+
 
 v0.51.0
 -------------------------------